[None][test] Consolidate base multimodal input-processor tests and register in CI
Rename test_input_processor_deterministic_math.py to
test_multimodal_input_processor.py and merge test_dummy_inputs_builder.py into
it, so a single file covers both base mixins a multimodal input processor
inherits: BaseMultimodalInputProcessor (token-counting math) and
BaseMultimodalDummyInputsBuilder (the modality-agnostic dummy-profiling
contract). Also repairs the stale _Sub stub (adds the call_with_text_prompt
abstract method).

Register the consolidated file in tests/integration/test_lists/test-db/l0_a10.yml
alongside the other inputs/ CPU unit tests so it is actually invoked in
pre-merge CI (neither original file was listed in any test-db, so they never
ran in CI).
